backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiosmtplib==3.0.1
email-validator==2.1.1
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
pdfminer.six==20221105
pytesseract==0.3.10
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
faker==24.11.0
aiosqlite==0.20.0
greenlet==3.0.3
